loggedin Header api changes
